requirements.txt
python-dotenv==0.8.2
tweepy==3.6.0
